// src/app/components/tms/model/accident.model.ts
export class AccidentModel {

    accidentExpenseId: number;
    vehicleId: number;
    //vehicleRegNumber:string;
    accidentDate: Date;
    accidentLocation: string;
    typeId: number;
    //accidentTypeName:string;
    involveWith: string;
    amountStatus: number;
    totalAmount: number;
    insuranceClaim: number;
    insuranceRecovered: number;
    budgetCodeId: number;
    odometer: number;
    status: number;
    remarks: string;

}